requirements.txt
fastapi
uvicorn[standard]
litellm
PyGithub
python-dotenv
pytest
coverage
